parasite: fix typecasting error
This patch fixes below typecasting error:
error: format ‘%lx’ expects argument of type ‘long unsigned int’, but argument 5 has type ‘uint64_t’ [-Werror=format]

Acked-by: Alexander Kartashov <alekskartashov@parallels.com>

+++ parasite-syscall.c
@@ -730,7 +730,8 @@ int parasite_fixup_vdso(struct parasite_ctl *ctl, pid_t pid,
 
 		off = (vma->vma.start / PAGE_SIZE) * sizeof(u64);
 		if (lseek(fd, off, SEEK_SET) != off) {
-			pr_perror("Failed to seek address %lx\n", vma->vma.start);
+			pr_perror("Failed to seek address %lx\n",
+				  (long unsigned int)vma->vma.start);
 			ret = -1;
 			goto err;
 		}
